move uam config and test to config/
